Improve xieite::math::approximatelyEqual

+++ include/xieite/geometry/getInterceptX.hpp
@@ -10,6 +10,6 @@ namespace xieite::geometry {
 	template<std::derived_from<xieite::geometry::LineLike> L>
 	[[nodiscard]]
 	constexpr std::optional<xieite::geometry::Point> getInterceptX(const L& lineLike) noexcept {
-		return xieite::math::approximatelyEqual(xieite::geometry::getSlope(lineLike), static_cast<long double>(0)) ? std::nullopt : std::optional(xieite::geometry::Point(lineLike.start.x - lineLike.start.y * (lineLike.end.x - lineLike.start.x) / (lineLike.end.y - lineLike.start.y), 0));
+		return xieite::math::approximatelyEqual(xieite::geometry::getSlope(lineLike), 0) ? std::nullopt : std::optional(xieite::geometry::Point(lineLike.start.x - lineLike.start.y * (lineLike.end.x - lineLike.start.x) / (lineLike.end.y - lineLike.start.y), 0));
 	}
 }
